+++ NEWS.md
@@ -1,6 +1,45 @@
+# colorspace 2.1-0
+
+* Bug fix for color vision deficiency simulations in `simulate_cvd()` based
+  on the work of [Machado _et al._ (2009)](https://doi.org/10.1109/TVCG.2009.113):
+  Following some illustrations from the supplementary materials, the transformations
+  in previous versions of the package had been applied to gamma-corrected sRGB
+  coordinates directly. However, the Machado _et al._ paper implicitly relies
+  on a linear RGB space (see page 1294, column 1) where their linear matrix
+  transformations for simulating color vision deficiencies are applied.
+  Therefore, a new argument `linear = TRUE` has been added to `simulate_cvd()`
+  (and hence in `deutan()`, `protan()`, and `tritan()`) that first maps the
+  provided colors to linearized RGB coordinates, applies the color vision
+  deficiency transformation, and then maps back to gamma-corrected sRGB
+  coordinates. Optionally, `linear = FALSE` can be used to restore the behavior
+  from previous versions where the transformations are applied directly to
+  the sRGB coordinates. For most colors the difference between the two
+  strategies is negligible but for some highly-saturated colors it becomes
+  more noticable, e.g., for red, purple, or orange. Thanks to Matthew Petroff
+  for reporting this issue and to Kenneth Knoblauch for advice and guidance.
+
+* Improvement in `simulate_cvd()` (and hence in `deutan()`, `protan()`, and
+  `tritan()`): When colors are specified as hex strings or named colors, the
+  sRGB coordinates after transformation are rounded appropriately to integers
+  in 0-255. Previous versions implicitly took the floor rather than round of
+  the coordinates.
+
+* Support formal S4 color objects in `simulate_cvd()` (and hence in `deutan()`,
+  `protan()`, and `tritan()`). In that case colors are transformed internally
+  to sRGB coordinates, color vision deficiency is simulated, and then a formal
+  S4 color object (of the same class as the input) is returned. This has the
+  advantage that no rounding is applied as when going through hex color strings.
+  Similarly, for an RGB matrix input, the output is also not rounded anymore.
+
+* Bug fix in `scale_colour_discrete_qualitative()` (and analogously for sequential,
+  diverging, and divergingx) to make sure that the scale functions can also
+  be called within a function whose arguments are controlled by an outer
+  function (reported by Marcelo S. Perlin).
+
+
 # colorspace 2.0-3
 
-* Added the advanced perceptual contrast algorithm (APCA) that is under
+* Added the accessible perceptual contrast algorithm (APCA) that is under
   development for the next major revision of the WCAG as an alternative
   `algorithm` in `contrast_ratio()`. See <https://github.com/Myndex/SAPC-APCA>.